+++ storage/client/blob/src/main/java/com/azure/storage/blob/BlobClient.java
@@ -39,7 +39,7 @@ import java.time.OffsetDateTime;
  *
  * <p>
  * This client contains operations on a blob. Operations on a container are available on {@link ContainerClient}, and
- * operations on the service are available on {@link StorageClient}.
+ * operations on the service are available on {@link BlobServiceClient}.
  *
  * <p>
  * Please refer to the <a href=https://docs.microsoft.com/en-us/rest/api/storageservices/understanding-block-blobs--append-blobs--and-page-blobs>Azure