+++ storage/client/blob/src/main/java/com/azure/storage/blob/models/ContainerSetAccessPolicyHeaders.java
@@ -5,6 +5,7 @@
 package com.azure.storage.blob.models;
 
 import com.azure.core.implementation.DateTimeRfc1123;
+import com.azure.core.implementation.annotation.Fluent;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
 import java.time.OffsetDateTime;
@@ -13,6 +14,7 @@ import java.time.OffsetDateTime;
  * Defines headers for SetAccessPolicy operation.
  */
 @JacksonXmlRootElement(localName = "Container-SetAccessPolicy-Headers")
+@Fluent
 public final class ContainerSetAccessPolicyHeaders {
     /*
      * The ETag contains a value that you can use to perform operations