+++ storage/client/blob/src/samples/java/com/azure/storage/blob/FileTransferExample.java
@@ -50,10 +50,10 @@ public class FileTransferExample {
         String endPoint = String.format(Locale.ROOT, "https://%s.blob.core.windows.net", accountName);
 
         /*
-         * Create a StorageClient object that wraps the service endpoint, credential and a request pipeline.
+         * Create a BlobServiceClient object that wraps the service endpoint, credential and a request pipeline.
          * Now you can use the storageClient to perform various container and blob operations.
          */
-        StorageClient storageClient = new StorageClientBuilder().endpoint(endPoint).credential(credential).buildClient();
+        BlobServiceClient storageClient = new BlobServiceClientBuilder().endpoint(endPoint).credential(credential).buildClient();
 
 
         /*