+++ storage/client/queue/src/main/java/com/azure/storage/queue/implementation/AzureQueueStorageBuilder.java
@@ -6,10 +6,12 @@ package com.azure.storage.queue.implementation;
 
 import com.azure.core.http.HttpPipeline;
 import com.azure.core.implementation.RestProxy;
+import com.azure.core.implementation.annotation.ServiceClientBuilder;
 
 /**
- * A appendBlobClientBuilder for creating a new instance of the AzureQueueStorage type.
+ * A builder for creating a new instance of the AzureQueueStorage type.
  */
+@ServiceClientBuilder(serviceClients = AzureQueueStorageImpl.class)
 public final class AzureQueueStorageBuilder {
     /*
      * The URL of the service account, queue or message that is the targe of the desired operation.
@@ -73,9 +75,11 @@ public final class AzureQueueStorageBuilder {
         }
         AzureQueueStorageImpl client = new AzureQueueStorageImpl(pipeline);
         if (this.url != null) {
-            client.url(this.url);
+            client.setUrl(this.url);
         }
-        client.version(this.version);
+        if (this.version != null) {
+            client.setVersion(this.version);
+        }
         return client;
     }
 }