+++ storage/client/queue/src/main/java/com/azure/storage/queue/models/Metrics.java
@@ -4,6 +4,7 @@
 
 package com.azure.storage.queue.models;
 
+import com.azure.core.implementation.annotation.Fluent;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
 
@@ -11,6 +12,7 @@ import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
  * The Metrics model.
  */
 @JacksonXmlRootElement(localName = "Metrics")
+@Fluent
 public final class Metrics {
     /*
      * The version of Storage Analytics to configure.